Configure default "Engineering" Shortcut workflow
If the user has a Shortcut Story Workflow named "Engineering", we
detect that, grab its ID, and populate a config.json file with a
"workflow_id" entry with that value.

If the user does not have a workflow named "Engineering", we detect
that, populate the config.json "workflow_id" with a null value, then
print out the workflows the user does have so they can select the
appropriate one and paste it into config.json themselves.