Progress toward generating data/states.csv
Handles the happy path of finding an "Engineering" workflow with state
names that are the default.

Handles the less happy path of not finding an "Engineering" workflow
in the user's workspace. In this case, the script prints out
instructions, the workflows in their workspace with states, and
instructs them to use the ID of the workflow in their config.json and
run it again.

Also handles the less happy path of the "Engineering" workflow being
identified, but one of the states not matching the expected names. In
this case, similar flow: instructions, all states printed, user
expected to manually update data/states.csv and run initialize.py
again to have it validated.

As of this commit, this script prints out a pre-canned message for
success and 0 of every kind of entity processed.