Add lint script and Makefile target; lint code base
I added the Black linter several commits back, but hadn't put an easy
facility into place to leverage it. This commit adds a script, a
Makefile target that runs it, and adds a Makefile target of setup-dev
which can be used instead of setup to include dependencies marked as
development-only in the Pipfile.