Generate data/users.csv to map Pivotal and Shortcut users
This handles the case where no data/users.csv is present. It generates
one based on all of the unique users found in the Pivotal export,
attempting to map automatically if user names match those found in the
Shortcut workspace.

If there are any users left unmapped, they are printed to the console
and a data/shortcut_users.csv is generated as a reference for all
users found in their Shortcut Workspace.

A future commit will handle the case of validating an existing
data/users.csv file and sharing those email addresses which are found
in it but not yet in the Shortcut Workspace.