Add pytest-cov to dev dependencies and use in test script
I've set this to fail if the code coverage drops below 30%. We're
currently at ~37%.

Obviously this is a low threshold. Our current focus is getting a
working import with more manual testing as we develop the
importer. Some refactoring and/or mocking of functions that make API
calls to Shortcut will allow us to make the code more friendly to
automated testing.