Pivotal Import, Alpha (#38)
* Initial design and file layout

* Create initial pipenv files, add requests as first dependency

* Remove states.csv, will generate using Shortcut data

* Add Black for automatic Python formatting

* Initial Shortcut API integration and CLI entry point

* Ignore CSV and JSON files produced by initialize.py

These files are intended to be generated by initialize.py and then
optionally edited by the end-user to define configuration values to
use during the import.

* Configure default "Engineering" Shortcut workflow

If the user has a Shortcut Story Workflow named "Engineering", we
detect that, grab its ID, and populate a config.json file with a
"workflow_id" entry with that value.

If the user does not have a workflow named "Engineering", we detect
that, populate the config.json "workflow_id" with a null value, then
print out the workflows the user does have so they can select the
appropriate one and paste it into config.json themselves.

* Make printing of Workflow + states clearer

* Ensure data directory is present

* Ignore data/pivotal_export.csv

* Progress toward generating data/states.csv

Handles the happy path of finding an "Engineering" workflow with state
names that are the default.

Handles the less happy path of not finding an "Engineering" workflow
in the user's workspace. In this case, the script prints out
instructions, the workflows in their workspace with states, and
instructs them to use the ID of the workflow in their config.json and
run it again.

Also handles the less happy path of the "Engineering" workflow being
identified, but one of the states not matching the expected names. In
this case, similar flow: instructions, all states printed, user
expected to manually update data/states.csv and run initialize.py
again to have it validated.

As of this commit, this script prints out a pre-canned message for
success and 0 of every kind of entity processed.

* Document functions

* When printing Shortcut workflows, also write to CSV

* Ignore Shortcut workflows written to CSV

* Extract functions for interacting with SC API

* Beginning of importer from CSV files

* Add logger to api and add some docstrings

* Select desired keys for importing stories

* Add Makefile to quickly cleanup generated files

* Deduce workflow when creating a story

* Fix bulk create

* Include story type

* Add created_at

* Print some import stats after the import is complete

* Add setup script for Python tooling

* Add setup Makefile target

* Make clean target exit with success even if files absent

* Add lint script and Makefile target; lint code base

I added the Black linter several commits back, but hadn't put an easy
facility into place to leverage it. This commit adds a script, a
Makefile target that runs it, and adds a Makefile target of setup-dev
which can be used instead of setup to include dependencies marked as
development-only in the Pipfile.

* Use /bin/sh instead of bash in scripts

* Add linting via GitHub Actions for pivotal-import project

* Update README, wording changes and instructions

Also to trigger the GitHub Actions for the first time.

* Support comments

* Cleanup type errors and warnings

* Import tasks

* Import estimate

* Import labels

* Pass in context

* First pass at importing epics

Does not link epics with stories
The stories contained in an epic share a label.

Followup work:
Add the stories to the appropriate epics

* Add pytest dep, test target in Makefile

* Extract parse_comment, parse_date for common use

Will need the parse_comment function to determine users found as
authors of comments. The parse_comment function relies on parse_date

* Run tests in GitHub Actions CI

* Permit multi-line comments

* Remove unused imports from initialize.py

* Add make target for running initialize.py

* Ignore data/shortcut_users.csv

* Generate data/users.csv to map Pivotal and Shortcut users

This handles the case where no data/users.csv is present. It generates
one based on all of the unique users found in the Pivotal export,
attempting to map automatically if user names match those found in the
Shortcut workspace.

If there are any users left unmapped, they are printed to the console
and a data/shortcut_users.csv is generated as a reference for all
users found in their Shortcut Workspace.

A future commit will handle the case of validating an existing
data/users.csv file and sharing those email addresses which are found
in it but not yet in the Shortcut Workspace.

* Making bulk operation more generic

* Create epics and then link them to their stories

Refactored importer slightly to allow processing of epics first

* Add tests for import steps

* Update doc to use new file name

* Some epic tests

* Identify unmapped, uninvited users

* Fuzzy matching users

Employs some heuristics to potentially find matches for users
- lower cases and removes spaces and underscores in the names
- tries to find a match based on the SC mention name
- uses a fuzzy search algorithm to get the closes match

* Moved this ignore to my global

* Map user names for import

* Fix label parsing

* Add pytest-cov to dev dependencies and use in test script

I've set this to fail if the code coverage drops below 30%. We're
currently at ~37%.

Obviously this is a low threshold. Our current focus is getting a
working import with more manual testing as we develop the
importer. Some refactoring and/or mocking of functions that make API
calls to Shortcut will allow us to make the code more friendly to
automated testing.

* Add an import Makefile target that depends on initialize

* Support -h/--help for initialize.py

* Update comments, remove unused code

* Remove unused imports from pivotal_import.py

* Edit code comments describing mapping of entity types

* Add flag to turn on debug logging

* Support multiple owners

* Actually finish remove the bulk option

* Update limitations as of this commit in the README

* Direct users to appropriate feedback channels

* Create GitHub issue templates for bug reports, feature requests

These are global to the useshortcut/api-cookbook repository.

These two templates were adapted from those found in the GitHub
appium/appium repository:

https://github.com/appium/appium/tree/59ef09d9558f1a6a9f4ee3d81d4cb0b7218706b4/.github/ISSUE_TEMPLATE

* Notice of alpha quality and Discord link

* Add notes concerning Shortcut billing, script usage

* Add Makefile target for import-apply

Also fixes bug in setup-dev target

---------

Co-authored-by: Christopher Demwell <chris.demwell@shortcut.com>
Co-authored-by: Osei Poku <opoku@users.noreply.github.com>
Co-authored-by: Osei Poku <osei.poku@shortcut.com>