[pivotal] Map Pivotal priority to Shortcut Priority custom field
This defaults to the built-in Priority custom field in the Shortcut
Workspace. That custom field has a canonical_name of priority; its
custom field values can have neither their position nor their icon
edited; they only permit changing the user-facing name that appears
for custom field values in Shortcut. This means that we can directly
map p0 through p3 from Pivotal to the custom field values of the
built-in Priority custom field with positions 0 through 3.

This implementation treats a Pivotal priority of `none` such that no
Shortcut Custom Field Value is set during import.

Otherwise this mapping experience mirrors what is done for Pivotal
story states. If the Priority field is enabled, it is used, otherwise
the user is prompted with all of the Custom Fields defined (and
enabled) in their workspace to select a target Custom Field. Once a
target Custom Field is configured in config.json, its Custom Field
Values are used to map Pivotal states in position order. If there are
any gaps in the mapping, an error message is presented to the user
with instructions on completing the mapping.