[pivotal] Import story priority as Priority custom field values
The Shortcut Priority Custom Field is enabled in all new workspaces by
default.

A previous commit introduced the initialization routines to populate a
local data/priorities.csv file that includes a mapping from Pivotal
priority to the Priority Custom Field in Shortcut.

This commit leverages that mapping to enhance imported Shortcut
Stories with a Priority custom field value.

Pivotal priority p0 through p3 are mapping to the Priority Custom
Field Values in Shortcut with positions 0 through 3 (commonly named
"Highest", "High", "Medium", and "Low", though users are free to
rename them). Pivotal priority of "none" is considered a signal not to
set the Priority field in Shortcut at all, so that value is ignored
during import.