[pivotal] Import Pivotal reviews as Shortcut comments
In a previous commit, this importer was extended to map all Pivotal
story reviewers as Shortcut story followers, so that they will receive
in-app notifications about activity on the stories for which they were
reviewers in Pivotal.

This commit goes a step further and captures the information about
reviewer, review type, and review status in a comment on the
story. There is explanatory text at the beginning of the comment
followed by a Markdown-formatted table for each review which renders
nicely in Shortcut.

This review-as-comment is created last and is thus the newest comment
on the imported story, which means that by default it renders last in
the UI (users can choose to order comments in either chronological
order).