[pivotal] Print deletion progress and failures to the console
The pivotal_import.py script is able to print the Shortcut label
app_url for the user during the normal course of the import.

As of this commit, we do not keep ready access to that app_url either
in a JSON or CSV file that we write to disk, so it's not readily
available to provide to the user during deletion.

So this commit includes a simple print statement that prints one line
per entity deleted, conveying both the type (story/epic) and the ID of
the Shortcut entity being deleted.

If a deletion request to the Shortcut API fails, this commit also
prints that error message to the console.