[pivotal] Try to detect default Team; prompt user if team not found.
NOTE: These changes affect _only_ the initialization part of the
importer. Future commits will need to leverage the initialized data to
set a team/group for imported Shortcut entities.

Teams are a central part of the Shortcut model, but users may or may
not have set up multiple teams when trying out this importer.

During initialization of the importer, these changes will search in
the user's Shortcut workspace for the default team named "Team 1". If
that team is found, it will be automatically set as the target team
during the import, meaning that all epics and stories imported will
have that Team assigned to them.

If the default team is not found, rather than printing out
"[Problem]", the script will print out "[Warning]" so that, if the
user desires, they can look at the teams/groups printed to the console
or consult the data/groups.csv file to pick a Team to use for the
import. If no team/group is configured in the `config.json` file, the
intention is that the importer will simply omit assigning a team to
imported epics and stories.