[pivotal] Assign imported epics/stories a Team/Group
NOTE: Shortcut's REST API uses the word "Group" to refer to the
concept that in Shortcut's UI and documentation is called "Team". A
Team is a group of Shortcut users; stories can belong to one Team;
epics can belong to multiple Teams; and other entities in the Shortcut
domain model can also have Team associations (e.g., Iterations,
Objectives). You should completely ignore any references to `team_id`
that you see in the Shortcut REST API; these refer to a deprecated
concept and are not referring to the Team/Group concept used here.

The config.json file must include a "group_id" entry which may be null
or the ID of a Shortcut Team/Group. This commit leverages that entry
to assign this Team/Group to every epic and story imported into
Shortcut.