[pivotal] Code comment that Pivotal epics don't have a requester
While working on tests in a previous commit, I noticed that we were
not providing a `requested_by_id` field in our Shortcut "Create Epic"
calls and thought this was an oversight.

Upon further inspection, Pivotal's domain model does not expose a
requester (or owners) for epics, and thus their export leaves those
entries blank as well.

In practice this means every imported epic is requested by the
importing user and has no owners.

The previous commit, however, does assign a Team/Group to imported
epics, so that is at least one way for users to group epics as they
import them.