[pivotal] Fix bug that assigned all stories to one epic
The Pivotal export relies on labels to demonstrate epic membership of
Pivotal stories. This importer follows a similar approach, whereby we
rely on a shared label between epics and stories to determine epic
membership.

A bug was introduced that erroneously mapped our timestamped labels to
an epic, thus marking every story being imported as belong to a
specific epic.

This commit corrects that behavior by ignoring that
PIVOTAL_TO_SHORTCUT_RUN_LABEL lable when processing the mapping of
epics and their labels.

This commit also refactors the code to make the essential act of
assigning an epic_id entry to stories being prepared for import
testable, and adds a test to demonstrate that epic_id is only assigned
if the story is in fact labeled with that epic.