[pivotal] Import the deadline field on (release) stories
Pivotal's export includes a Deadline entry for Pivotal Releases. This
importer was already importing releases as Shortcut stories, setting
their Shortcut story type to "chore", and labelling them as a
release. This commit ensures the deadline field is also imported and
set as the Shortcut Story's deadline.

Note that Shortcut stories of any type can have a deadline set, not
just releases, but Pivotal supports deadlines only on releases.