[pivotal] Support import and deletion of Iterations
The rows of the Pivotal export CSV each represent either a story or an
epic. The Pivotal iteration to which a story might belong is found in
an Iteration column, and when that is defined there are also populated
Iteration Start and Iteration End columns.

This commit extends the build_entity function to return an iteration
and pt_iteration_id entry for every row. The iteration entry is a
string containing the id, start, and end dates. A string was chosen
instead of a dict or other data structure, because those are not
hashable and thus not includable in a Python set, whereas a string
is. At commit() time, the unique set of these iteration strings is
used to create Shortcut Iterations.

The title of imported Iterations is "PT {id}" where id is the Pivotal
iteration ID. The start and end dates are also imported. At this time,
no Team/Group association is made for imported Iterations.

At commit() time this code also ensures each story is associated with
the correct imported Iteration.