[pivotal] Add function to perform file uploads via Shortcut API
The new `sc_upload_files` function takes a Shortcut story ID and a
list of file paths, uploads the files to Shortcut, and then associates
the files with the given story.

The Shortcut API allows for uploading up to 4 files in a single
request, but given lack of retry logic, size checking, etc., I have
opted in this commit to upload one file at a time and _not_ to stop
execution if an error is raised. A warning is printed to the console
if a file could not be uploaded for whatever reason.