[pivotal] Support import and deletion of files
This commit wires up the `sc_upload_files` function introduced in a
previous commit.

The Pivotal CSV export includes folders whose names are Pivotal story
IDs. Each folder contains all of the files uploaded to Pivotal Tracker
that were associated with that Pivotal story.

This importer now detects the presence of those folders inside the
`data/` folder of this project. If present, files inside of it are
uploaded to Shortcut, and then associated with stories as the stories
are imported.

The deletion script included with this importer has been updated to
delete files listed in the shortcut_imported_entities.csv file.