Update set-up-instructions.md
Changed guidance to use zshenv since zsh has been default shell for MacOS since the Catalina release in Oct 2019.

Added the expected output of `echo $SHORTCUT_API_TOKEN`. If the variable isn't exported correctly the response is blank, but there is no error. This could be interpreted as expected since this method is promoted as being more secure.