post-import-utils/inject_blockers.py: Stop skipping resolved blockers.
I had originally skipped resolved blockers because I didn't know how to represent the resolution state in Shortcut. Now I know (and like) that in Shortcut, a blocker has no resolution state of its own; rather, Shortcut infers the state of the blocker from the state of the blocking story.